fix build for non-x86 arch builds
